refactor: move editor-related components and hooks to dedicated directories
