+++ app/components/hooks/useShareInfo.tsx
